back-end-python/pyproject.toml
[project]
name = "langchain-rag-demo"
version = "0.1.0"
description = "LangChain RAG Demo - 环境小助手智能问答系统"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your@email.com" }]
keywords = ["langchain", "rag", "llm", "ai"]

dependencies = [
    # LangChain 核心
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-text-splitters>=0.3.0",

    # 阿里云通义千问
    "dashscope>=1.20.0",

    # 向量数据库
    "chromadb>=0.5.0",

    # 文档处理
    "pypdf>=4.0.0",

    # Web 应用
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-limiter>=3.0.0",

    # 环境变量
    "python-dotenv>=1.0.0",

    # 中文文本处理（BM25需要）
    "jieba>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

langgraph = [
    "langgraph>=0.2.0",
]

[project.scripts]
rag-build = "main:build_knowledge_base"
rag-ask = "main:ask_question"
rag-chat = "main:interactive_mode"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
pythonpath = ["."]
markers = [
    "integration: tests that require model API and external services"
]